+++ lib/Convos/Command/backend.pm
@@ -2,17 +2,16 @@ package Convos::Command::backend;
 
 =head1 NAME
 
-Convos::Command::backend - Start convos backend
+Convos::Command::backend - Start Convos backend
 
 =head1 DESCRIPTION
 
-This convos command will start the convos backend in a standalone process.
-At the same time a lock file is created preventing a frontend from starting
-the backend.
+This command is used to start the Convos backend.
 
 =cut
 
 use Mojo::Base 'Mojolicious::Command';
+use Daemon::Control;
 
 =head1 ATTRIBUTES
 
@@ -26,10 +25,43 @@ Returns a string describing how to use this command.
 
 =cut
 
-has description => "Start the Convos IRC proxy.\n";
+has description => "Start the Convos connection backend.\n";
 has usage       => <<"EOF";
-Usage: $0 backend
+Usage:
+
+# Fork to background
+$0 backend start
+
+# Control the forked backend
+$0 backend {restart|stop|status}
+
+# Run in foreground
+$0 backend
 EOF
+
+has _daemon => sub {
+  my $self = shift;
+  my $program = File::Spec->catfile($self->app->home, qw( script convos-backend ));
+
+  return Daemon::Control->new(
+    fork        => 2,
+    init_config => $ENV{CONVOS_INIT_CONFIG_FILE} || '/etc/default/convos',
+    init_code   => "MOJO_MODE='$ENV{MOJO_MODE}';",
+    help        => $self->usage,
+    name        => 'Convos backend',
+    lsb_desc    => $self->description,
+    lsb_sdesc   => 'Convos backend',
+
+    directory => $self->app->home,
+    program   => -x $program ? $program : 'convos-backend',
+    pid_file  => $ENV{CONVOS_BACKEND_PID_FILE} || File::Spec->catfile(File::Spec->tmpdir, 'convos-backend.pid'),
+    stderr_file => $ENV{CONVOS_BACKEND_LOGFILE} || File::Spec->devnull,
+    stdout_file => $ENV{CONVOS_BACKEND_LOGFILE} || File::Spec->devnull,
+
+    user  => $ENV{RUN_AS_USER},
+    group => $ENV{RUN_AS_GROUP},
+  );
+};
 
 =head1 METHODS
 
@@ -41,27 +73,37 @@ Will start the backend.
 
 sub run {
   my ($self, @args) = @_;
-  my $app   = $self->app;
-  my $loop  = Mojo::IOLoop->singleton;
-  my $redis = $app->redis;
+  my $daemon = $self->_daemon;
 
-  unless ($ENV{MOJO_MODE}) {
-    die qq(MOJO_MODE need to be set to either "production" or "development".\n);
+  $daemon->read_pid;
+
+  if (@args) {
+    local $ENV{CONVOS_IS_CONTROLLED_BY_DC} = 1;
+    $self->_exit($daemon->run_command(@args));
+  }
+  if (!$ENV{CONVOS_IS_CONTROLLED_BY_DC} and $daemon->pid and $daemon->pid_running) {
+    $self->app->log->warn('Backend is already running.');
+    return 0;
   }
 
-  $SIG{QUIT} = sub {
-    $app->log->info('Gracefully stopping backend...');
-    $loop->max_connnections(0);
-    $redis->del('convos:backend:pid') if $redis;
-  };
+  $0 = "$0-backend" unless $0 =~ /-backend$/;    # easier to spot the process
+  $daemon->pid($$);
+  $daemon->write_pid;
+  $self->{running} = 1;
+  $self->app->log->info('Starting convos backend.');
+  $self->app->core->start;
+  Mojo::IOLoop->start unless $ENV{CONVOS_BACKEND_EMBEDDED};
+}
 
-  $ENV{CONVOS_BACKEND_EMBEDDED} = 1;
-  $redis->set('convos:backend:pid', $$);
-  $redis->expire('convos:backend:lock' => 1);
-  $app->core->start;
-  $app->log->info('Starting convos backend');
-  $loop->start;
-  return 0;
+sub DESTROY {
+  my $self = shift;
+  my $file = $self->_daemon->pid_file;
+
+  unlink $file if $self->{running} and -w $file;
+}
+
+sub _exit {
+  exit $_[1];
 }
 
 =head1 COPYRIGHT