+++ t/Helper.pm
@@ -8,9 +8,8 @@ use Mojo::Redis;
 BEGIN {
   $ENV{MOJO_MODE} = 'testing';
   $ENV{CONVOS_ARCHIVE_DIR} //= 'irc_logs_test';
-  $ENV{CONVOS_DEBUG} //= $ENV{TEST_VERBOSE};
-  $ENV{CONVOS_MANUAL_BACKEND} = 1;
-  $ENV{CONVOS_REDIS_URL} //= '';
+  $ENV{CONVOS_DEBUG}       //= $ENV{TEST_VERBOSE};
+  $ENV{CONVOS_REDIS_URL}   //= '';
 }
 
 my ($redis, $t);